+++ src/Infrastructure/Dto/Order/OrderResponse.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Domain\Order\Dto;
+namespace App\Infrastructure\Dto\Order;
 
 use App\Domain\Order\Entity\Order;
 use App\Domain\Thing\Dto\ThingOut;
@@ -10,7 +10,7 @@ use App\Domain\User\Dto\RequesterOut;
 use Nelmio\ApiDocBundle\Annotation\Model;
 use Swagger\Annotations as SWG;
 
-class OrderOut
+class OrderResponse
 {
     /** @SWG\Property(type="string") */
     public string $id;
